requirements.txt
Django>=3.2.5, <3.3.0
djangorestframework>=3.12.4, <3.13.0
psycopg2>=2.9.1, <2.10.0
flake8>=3.9.2, <3.10.0
Pillow>=8.3.1, <8.4.0
django-phonenumber-field>=5.2.0, <5.3.0
phonenumbers>=8.12.27, <8.13.0
python-decouple==3.4
gunicorn==20.1.0
django-heroku==0.3.1
django-storages==1.11.1
boto3>=1.18.6, <1.19.0
django-money>=2.0.1, <2.1.0
PyYAML
django-bootstrap4